+++ Sources/StreamChat/Database/DTOs/ChannelDTO.swift
@@ -65,16 +65,8 @@ class ChannelDTO: NSManagedObject {
     override func willSave() {
         super.willSave()
 
-        // Change to the `truncatedAt` value have effect on messages, we need to mark them dirty manually
-        // to triggers related FRC updates
-        if changedValues().keys.contains("truncatedAt") {
-            messages
-                .filter { !$0.hasChanges }
-                .forEach {
-                    // Simulate an update
-                    $0.willChangeValue(for: \.id)
-                    $0.didChangeValue(for: \.id)
-                }
+        guard !isDeleted else {
+            return
         }
         
         // Update the date for sorting every time new message in this channel arrive.