+++ Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -69,14 +69,18 @@ class MessageDTO: NSManagedObject {
     
     override func willSave() {
         super.willSave()
+
+        guard !isDeleted else {
+            return
+        }
         
         // Manually mark the channel as dirty to trigger the entity update and give the UI a chance
         // to reload the channel cell to reflect the updated preview.
-        if let channel = previewOfChannel, !channel.hasChanges, isUpdated {
+        if let channel = previewOfChannel, !channel.hasChanges, !channel.isDeleted {
             let cid = channel.cid
             channel.cid = cid
         }
-        
+
         prepareDefaultSortKeyIfNeeded()
     }