+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # StreamChat iOS SDK CHANGELOG
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-# Upcoming
-
-### 🔄 Changed
+## StreamChat
+### ✅ Added
+- Add hide history option when adding a new member [#2155](https://github.com/GetStream/stream-chat-swift/issues/2155)
+### 🐞 Fixed
+- Avoid triggering CoreData updates in willSave() [#2156](https://github.com/GetStream/stream-chat-swift/pull/2156)
+- Fix Channel missing messages from NSE push updates [#2166](https://github.com/GetStream/stream-chat-swift/pull/2166)
+- Sync active channels when no channel list [#2163](https://github.com/GetStream/stream-chat-swift/pull/2163)
 
 # [4.18.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.18.0)
 _July 05, 2022_